disable keyboard popup onnexport (#112)
* more inputmode

this time, set it to none to prevent the keyboard from popping up. since
there's only a few options, this is ok

reasoning: I tapped the button to select 750 instead of 250, and the
keyboard pop-up caused the entire page to scroll just as I tapped the 750;
instead I tapped on a diaper link I didn't intend to, marking the 7th (or
whatever number) section as downloaded. frustratingly, the 7th(?) download
remained inaccessible even after I changed the number, so I had to reload
the page...and very nearly did it again

* *all* sliders only take numeric inputs

thats why thwyre sliders and not drop-downs...