More refactoring + cleanup
Now, the metadata is added as soon as the image is downloaded from the
server. Then, if the image is saved again in a different format, we just
grab the exif data from the old blob and apply it to the new one.

This makes for cleaner code, and prevents us from having to generate
the metadata info in separate places in the code